pyproject.toml
[project]
name = "tbi-lens"
version = "0.1.0"
description = "TBI-LENS: Traumatic Brain Injury Lesion Segmentation inference pipeline"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Sakshi Rathi", email = "rathi036@umn.edu" }]
license = { text = "MIT" }
dependencies = [
  "monai>=1.3.0",
  "torch>=2.0.0",
  "pyyaml>=6.0",
  "nibabel>=5.1.0",
  "numpy>=1.23.0",
]

[project.scripts]
tbi-lens = "tbi_lens.cli:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"
